rename page component
